add sweet alert

+++ src/app/components/product/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, RouterLink } from '@angular/router';
 import { ProductService } from '../../../services/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -34,17 +35,31 @@ export class ProductListComponent implements OnInit{
   }
 
   // Delete Product
-  DeleteProduct(id: any) {
-    this.productService.Delete(id).subscribe({
-      next: (res) => {
-        console.log(res);
-        this.ngOnInit();
-      },
-      error: (e) => {
-        console.log(e);
-        this.ngOnInit();
-      },
-    });
-  }
+
+DeleteProduct(id: any) {
+  Swal.fire({
+    title: 'Are you sure?',
+    text: 'This product will be permanently deleted!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#6c757d',
+    confirmButtonText: 'Yes, delete it!',
+    cancelButtonText: 'Cancel'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.productService.Delete(id).subscribe({
+        next: (res) => {
+          Swal.fire('Deleted!', 'Product has been deleted.', 'success');
+          this.ngOnInit();
+        },
+        error: (err) => {
+          Swal.fire('Error!', 'Something went wrong.', 'error');
+          console.error(err);
+        }
+      });
+    }
+  });
+}
 
 }
