+++ src/app/components/supplier/supplier-list/supplier-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { SupplierService } from '../../../services/supplier.service';
 import { Router, RouterLink } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-supplier-list',
@@ -34,17 +35,40 @@ export class SupplierListComponent implements OnInit{
   }
 
   // Delete Supplier
-  DeleteSupplier(id: any) {
-    this.supplierService.Delete(id).subscribe({
-      next: (res) => {
-        console.log(res);
-        this.ngOnInit();
-      },
-      error: (e) => {
-        console.log(e);
-        this.ngOnInit();
-      },
-    });
-  }
+DeleteSupplier(id: any) {
+  Swal.fire({
+    title: 'Are you sure?',
+    text: 'This supplier will be permanently deleted!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#3085d6',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.supplierService.Delete(id).subscribe({
+        next: (res) => {
+          Swal.fire({
+            icon: 'success',
+            title: 'Deleted!',
+            text: 'Supplier has been deleted.',
+            timer: 2000,
+            showConfirmButton: false
+          });
+          this.ngOnInit();
+        },
+        error: (e) => {
+          console.log(e);
+          Swal.fire({
+            icon: 'error',
+            title: 'Error!',
+            text: 'Failed to delete supplier.'
+          });
+          this.ngOnInit();
+        },
+      });
+    }
+  });
+}
 
 }
